Island 'Tube Case': WIP - store 'innercuts' in islands.
This is the easiest part of the work - find and store edges that are borders
(seams for UVs e.g.), and yet 'inside' islands, i.e. used by two or more
polgons of the island.

Now, we have to detect dest polys that get mapped 'accross' such edges - this is
the real difficult part...

Note: not sure yet whether we'll need also all 'border' edges of an island,
easy to add later if needed anyway.